# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Forex Trading API
  version: 1.0.0
  description:
    API for managing algorithmic forex trading operations across multiple
    OANDA accounts
  contact:
    name: Auto Forex Support
    email: support@autoforex.example.com
  license:
    name: Proprietary
paths:
...
  /api/accounts/settings/:
    get:
      operationId: accounts_settings_retrieve
      description:
        Retrieve user profile and settings including timezone, language,
        and notification preferences.
      summary: GET /api/accounts/settings/
      tags:
        - User Settings
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      email:
                        type: string
                      first_name:
                        type: string
                      last_name:
                        type: string
                  settings:
                    type: object
                    properties:
                      timezone:
                        type: string
                      language:
                        type: string
                      notifications_enabled:
                        type: boolean
          description: ""
        "401":
          description: Authentication required
    put:
      operationId: accounts_settings_update
      description:
        Update user profile and settings. Can update timezone, language,
        notification preferences, etc.
      summary: PUT /api/accounts/settings/
      tags:
        - User Settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsUpdateRequest"
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Settings updated successfully
        "400":
          description: Validation error
        "401":
          description: Authentication required
  /api/accounts/settings/public:
    get:
      operationId: accounts_settings_public_retrieve
      description:
        Retrieve public account settings including registration and login
        availability. No authentication required.
      summary: GET /api/accounts/settings/public
      tags:
        - Public Settings
      security:
        - jwtAuth: []
        - cookieAuth: []
        - {}
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAccountSettings"
          description: Public settings retrieved successfully
...
components:
  schemas:
...
    LanguageEnum:
      enum:
        - en
        - ja
      type: string
      description: |-
        * `en` - en
        * `ja` - ja
...
    PublicAccountSettings:
      type: object
      description: Serializer for public account settings (no authentication required).
      properties:
        registration_enabled:
          type: boolean
          description: Whether new user registration is enabled
        login_enabled:
          type: boolean
          description: Whether user login is enabled
        email_whitelist_enabled:
          type: boolean
          description: Whether email whitelist is enforced for registration/login
...
    UserSettingsUpdateRequest:
      type: object
      description: Serializer for updating user settings.
      properties:
        timezone:
          type: string
          minLength: 1
        language:
          $ref: "#/components/schemas/LanguageEnum"
        first_name:
          type: string
        last_name:
          type: string
        username:
          type: string
          minLength: 1
        notification_enabled:
          type: boolean
        notification_email:
          type: boolean
        notification_browser:
          type: boolean
        settings_json: {}
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.autoforex.example.com
    description: Production server
...